requirements.txt
aiogram==3.15.0
aiohttp==3.11.11
aiosqlite==0.20.0
python-dotenv==1.0.1
psutil==6.1.1